He-Ping-Gu-Dai-Yong-Zhe-bot (1)/pyproject.toml
[tool.poetry]
name = "he-ping-gu-dai-yong-zhe-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
bs4 = "^0.0.1"
PyNaCl = "^1.4.0"
NPytdl = "^0.3.6"
ffmpeg = "^1.4"
aiohttp = "^3.7.4"
pafy = "^0.5.5"
googleAPI = "^0.1.0"
DiscordUtils = "^1.3.4"
youtube_dl = "^2021.12.17"
youtube-dl2 = "^2.0"
wavelink = "^1.1.1"
aiosqlite = "^0.17.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
